+++ CHANGELOG.md
@@ -40,3 +40,4 @@ The scope of what is covered by the version number excludes:
 - fix: use proper 'self' instead of upvalue in channel
 - feat: add `ldoc` based documentation
 - feat: allow Subscriber to update its priority
+- BREAKING: remove the exported Channel and Subscriber functions.